Create Cardano (ADA) cryptocurrency holders are eyeing this coin
Cardano (ADA) cryptocurrency holders are eyeing this coin 

Recently had a meeting with some friends of mine. We discussed popular crypto projects and what else might shoot up.  My companions keep a lot of their money in the Cardano (ADA) project. And I was surprised to note that at the moment they are actively looking at the Crypto Raccon (RoonX) project.

So far, the project is in the development stage, but they say that by mid-November, it should be launched. 
Briefly, what is Crypto Raccon (RoonX). It is the world's first decentralized lottery based on blockchain and regulated by AI. I think it is clear for everyone the development prospects of this project. 
Plus there are strong rumors about the companies that finance this project. We discussed with friends the advantages of the RoonX project. It's worth stopping here separately, I'll give more details tomorrow.