Create Bitcoin $10 million dollars, and two more coins that will grow hundreds of times!
Bitcoin $10 million dollars, and two more coins that will grow hundreds of times!

Chris Bellance lead analyst at Goldman Sachs Group assessed the prospects for cryptocurrencies over the next 5 years. In his opinion, most government funds will make a sharp purchase of BTC, ADA, ROONX. 

The main reason that will influence this decision is the destabilization of currency-dollar relations around the world and a possible crisis in the United States. Settlement in national currencies is currently a very complicated procedure. That is why government funds are now looking at alternative currencies. If the expert community understands the motivation to buy BTC and ADA, ROONX is a dark lot, which is just starting its way. But already ROONX is attracting the attention of a large number of users. Observe