Create Crypto  Raccoon (ROONX)
I often write about this project. But it's
definitely worth it. First of all, you should pay attention to the team and funds,
that are circling around this project. ROONX is still holding out and not attracting
large funds, their task is maximum decentralization of the project's assets. In general
ROONX team reminds me of a crypto project in the early days of this sphere. 

In their project they put the fundamental principles of the crypto world:
Decentralization and Anonymity. 

The project it self which is based on lottery, looks very promising.  The lottery market in 2024 amounted to 234 billion dollars, and the number of players is approaching 1 trillion. I guess
that's the blue ocean?